+++ js/modules/CounterMeasure.js
@@ -4,7 +4,7 @@ import {DEFENSE_COMPONENT_TYPES} from "./Constants.js";
 export class CounterMeasure extends DefenseComponent {
   /**
    * @param {string} name
-   * @param {number} probability
+   * @param {Fraction} probability
    * @param {boolean} guided
    */
   constructor(name, probability, guided) {
@@ -18,6 +18,7 @@ export class CounterMeasure extends DefenseComponent {
    * @return {number}
    */
   reduceAttackDamage(incomingDamage , attackingWeapon) {
-    return (attackingWeapon.isGuided === this.guided && Math.random() < this.probability) ? 0 : incomingDamage;
+    return (attackingWeapon.isGuided === this.guided && Math.random() < this.probability.toDecimal())
+      ? 0 : incomingDamage;
   }
 }
